refactor: Rename `…/event-detection/event-types/apiGateway.js`
