+++ sdk-js/src/lib/event-detection/event-types/api-gateway.js
